# tree: Improve performance of tree.AsJSON