--- a/pkg/util/strutil/BUILD.bazel
+++ b/pkg/util/strutil/BUILD.bazel
@@ -1,0 +1,19 @@
+load("//build/bazelutil/unused_checker:unused.bzl", "get_x_data")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "strutil",
+    srcs = ["util.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/util/strutil",
+    visibility = ["//visibility:public"],
+)
+
+go_test(
+    name = "strutil_test",
+    srcs = ["util_test.go"],
+    args = ["-test.timeout=295s"],
+    embed = [":strutil"],
+    deps = ["@com_github_stretchr_testify//require"],
+)
+
+get_x_data(name = "get_x_data")