@miretskiy 2022-09-14
Full benchmark stat:
[json_delta.txt](https://github.com/cockroachdb/cockroach/files/9570042/json_delta.txt)